chore: restructure codebase to improve modularity and maintainability
