+++ acses-website/src/components/home/hero-section/StackedImages.jsx
