+++ acses-website/src/components/home/statistics/Statistics.jsx
