+++ acses-website/src/layouts/RouteLayout.jsx
@@ -1,7 +1,7 @@
 import "./routeLayout.css";
 import { Outlet } from "react-router-dom";
-import NavBar from "../components/navbar/NavBar";
-import Footer from "../components/footer/Footer";
+import NavBar from "../components/shared/navbar/NavBar";
+import Footer from "../components/shared/footer/Footer";
 
 const RouteLayout = () => {
   return (
