+++ acses-website/src/pages/executives/Executives.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import NavBar from "../../components/navbar/NavBar";
-import Footer from "../../components/footer/Footer";
+import NavBar from "../../components/shared/navbar/NavBar";
+import Footer from "../../components/shared/footer/Footer";
 import executives from "./executiveList";
 
 const Executives = () => {
